FEAT(player hand): overhaul data flow
Move the player hand from the Player struct to the Hand and pass it
down to the Trick. The Trick passes the Player's hand to the Player,
who decides which Card to play. The Player passes back the Card
selected for play and the new hand without the played Card.